+++ src/collections/btree_map.rs
@@ -5,7 +5,7 @@
 use std::collections::BTreeMap;
 
 use iter::*;
-use iter::internal::*;
+use iter::plumbing::*;
 
 use vec;
