+++ src/collections/hash_set.rs
@@ -6,7 +6,7 @@ use std::collections::HashSet;
 use std::hash::{Hash, BuildHasher};
 
 use iter::*;
-use iter::internal::*;
+use iter::plumbing::*;
 
 use vec;
