+++ src/iter/enumerate.rs
@@ -1,4 +1,4 @@
-use super::internal::*;
+use super::plumbing::*;
 use super::*;
 use std::iter;
 use std::ops::Range;
