+++ src/iter/filter_map.rs
@@ -1,4 +1,4 @@
-use super::internal::*;
+use super::plumbing::*;
 use super::*;
 
 use std::fmt::{self, Debug};
