+++ src/iter/fold.rs
@@ -1,4 +1,4 @@
-use super::internal::*;
+use super::plumbing::*;
 use super::*;
 
 use std::fmt::{self, Debug};
