+++ src/iter/interleave_shortest.rs
@@ -1,4 +1,4 @@
-use super::internal::*;
+use super::plumbing::*;
 use super::*;
 
 /// `InterleaveShortest` is an iterator that works similarly to
