+++ src/iter/plumbing.rs
