+++ src/iter/unzip.rs
@@ -1,4 +1,4 @@
-use super::internal::*;
+use super::plumbing::*;
 use super::*;
 
 /// This trait abstracts the different ways we can "unzip" one parallel
